Fix package namespacing
