--- pulpcore/cli/__init__.py
@@ -1,11 +1,0 @@
-import pkg_resources
-
-from pulpcore.cli.common import main  # noqa: F401
-
-
-# Load plugins
-# https://packaging.python.org/guides/creating-and-discovering-plugins/#using-package-metadata
-discovered_plugins = {
-    entry_point.name: entry_point.load()
-    for entry_point in pkg_resources.iter_entry_points("pulp_cli.plugins")
-}
