+++ pulpcore/cli/common/openapi.py
